#65 split lexer unit tests into separate partial class files by token type
