+++ src/Kingsland.MofParser.UnitTests/CodeGen/MofGeneratorTests.cs
@@ -1,4 +1,5 @@
 ﻿using Kingsland.MofParser.CodeGen;
+using Kingsland.MofParser.Lexing;
 using Kingsland.MofParser.Parsing;
 using Kingsland.ParseFx.Parsing;
 using Kingsland.ParseFx.Text;
@@ -89,7 +90,7 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "\t[read: ToSubClass, MappingStrings{\"Win32API|AccessControl|Windows NT Privileges\"}: ToSubClass] string PrivilegesNotHeld[];\r\n" +
                     "\t[read: ToSubClass, MappingStrings{\"Win32API|AccessControl|Windows NT Privileges\"}: ToSubClass] string PrivilegesRequired[];\r\n" +
                     "};";
-                var tokens = Lexing.Lexer.Lex(SourceReader.From(sourceMof));
+                var tokens = Lexer.Lex(SourceReader.From(sourceMof));
                 var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
                 var ex = Assert.Throws<UnexpectedTokenException>(
                     () =>
@@ -172,7 +173,7 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "{\r\n" +
                     "\t100\r\n" +
                     "};";
-                var tokens = Lexing.Lexer.Lex(SourceReader.From(sourceMof));
+                var tokens = Lexer.Lex(SourceReader.From(sourceMof));
                 var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
                 Assert.AreEqual(sourceMof, tokensMof);
                 var ex = Assert.Throws<UnexpectedTokenException>(
@@ -305,7 +306,7 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "{\r\n" +
                     "\t100\r\n" +
                     "};";
-                var tokens = Lexing.Lexer.Lex(SourceReader.From(sourceMof));
+                var tokens = Lexer.Lex(SourceReader.From(sourceMof));
                 var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
                 var ex = Assert.Throws<UnexpectedTokenException>(
                     () => {
@@ -473,7 +474,7 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "{\r\n" +
                     "\tJuly = \"July\"\r\n" +
                     "};";
-                var tokens = Lexing.Lexer.Lex(SourceReader.From(sourceMof));
+                var tokens = Lexer.Lex(SourceReader.From(sourceMof));
                 var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
                 var ex = Assert.Throws<UnexpectedTokenException>(
                     () =>
@@ -1390,7 +1391,7 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
                     "{\r\n" +
                     "\tMonth = {MonthEnums.July};\r\n" +
                     "};";
-                var tokens = Lexing.Lexer.Lex(SourceReader.From(sourceMof));
+                var tokens = Lexer.Lex(SourceReader.From(sourceMof));
                 var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
                 var ex = Assert.Throws<UnexpectedTokenException>(
                     () =>
@@ -1449,7 +1450,7 @@ namespace Kingsland.MofParser.UnitTests.CodeGen
         private static void AssertRoundtrip(string sourceMof, ParserQuirks parserQuirks = ParserQuirks.None)
         {
             // check the lexer tokens roundtrips ok
-            var tokens = Lexing.Lexer.Lex(SourceReader.From(sourceMof));
+            var tokens = Lexer.Lex(SourceReader.From(sourceMof));
             var tokensMof = TokenMofGenerator.ConvertToMof(tokens);
             Assert.AreEqual(sourceMof, tokensMof);
             // check the parser ast roundtrips ok
