+++ src/Kingsland.MofParser.UnitTests/Lexing/TestCases/MyServer.mof
